Renamed JsonPreviewer to PreviewContainer
