+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { MantineProvider } from "@mantine/core";
-import PreviewContainer from "./JsonPreviewer";
+import PreviewContainer from "./PreviewContainer";
 import "./App.css";
 import "@mantine/core/styles.css";
 import "@mantine/tiptap/styles.css";
